+++ locales/it/index.yml
@@ -3808,6 +3808,8 @@ permissionRequest:
     cta: Apri Impostazioni
 FIMS: 
   loadingScreen:
+    abort:
+      title: "Operazione annullata"
     consents:
       title: "Attendi qualche secondo"
     in-app-browser:
