+++ ts/features/fims/store/reducers/index.ts
@@ -1,15 +1,18 @@
 import * as pot from "@pagopa/ts-commons/lib/pot";
+import { pipe } from "fp-ts/lib/function";
+import * as O from "fp-ts/lib/Option";
 import { getType } from "typesafe-actions";
 import { Action } from "../../../../store/actions/types";
 import { GlobalState } from "../../../../store/reducers/types";
 import {
+  fimsCancelOrAbortAction,
   fimsGetConsentsListAction,
   fimsGetRedirectUrlAndOpenIABAction
 } from "../actions";
 import { ConsentData } from "../../types";
-import { isStrictNone } from "../../../../utils/pot";
+import { foldK, isStrictNone } from "../../../../utils/pot";
 
-type FimsFlowStateTags = "consents" | "in-app-browser";
+type FimsFlowStateTags = "consents" | "in-app-browser" | "abort";
 
 export type FimsState = {
   currentFlowState: FimsFlowStateTags;
@@ -52,12 +55,41 @@ const reducer = (
         ...state,
         consentsData: pot.toError(state.consentsData, action.payload)
       };
+    case getType(fimsCancelOrAbortAction):
+      return pipe(
+        state.consentsData,
+        abortUrlFromConsentsPot,
+        O.fold(
+          () => ({ ...state }),
+          () => ({ ...state, currentFlowState: "abort" })
+        )
+      );
   }
   return state;
 };
 
 export const fimsConsentsDataSelector = (state: GlobalState) =>
   state.features.fims.consentsData;
+
+export const fimsPartialAbortUrl = (state: GlobalState) =>
+  pipe(state, fimsConsentsDataSelector, abortUrlFromConsentsPot, O.toUndefined);
+
+const abortUrlFromConsentsPot = (consentsPot: pot.Pot<ConsentData, string>) =>
+  pipe(
+    consentsPot,
+    foldK(
+      () => O.none,
+      () => O.none,
+      newConsents => O.some(newConsents),
+      _ => O.none,
+      consents => O.some(consents),
+      consents => O.some(consents),
+      (_, newConsents) => O.some(newConsents),
+      (consents, _) => O.some(consents)
+    ),
+    // eslint-disable-next-line no-underscore-dangle
+    O.map(consents => consents._links.abort.href)
+  );
 
 export const fimsErrorStateSelector = (state: GlobalState) =>
   // this selector will be used to map the error message
